back-end/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
Django==5.0.4
djangorestframework==3.15.1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
psycopg==3.1.18
psycopg-binary==3.1.18
python-dotenv==1.0.1
slack_sdk==3.27.1
slackclient==2.9.4
sqlparse==0.4.4
typing_extensions==4.11.0
yarl==1.9.4
